Throw JS stuff into JSVyxal because why not
