+++ index.html
@@ -94,7 +94,7 @@
     </script>
 </head>
 
-<body onload="initCodeMirror(); decodeURL();  updateCount();">
+<body onload="Vyxal.initCodeMirror(); Vyxal.decodeURL(); Vyxal.updateCount();">
     <textarea id=dummy style="visibility: hidden;position:absolute; height: 0px; font: size 20px;"></textarea>
     <session-code>{{session}}</session-code>
     <h2 style="display: inline-block;"><a href="https://github.com/Vyxal/Vyxal" target="_blank">Vyxal</a></h2>
